Github comments
Renamed files. Renamed Graph class to CheckGraph. Updated type hints. Added more docs to classes. Moved lambda function pricing logic from pricing.py to lambda_function_pricing.py. Moved various ask(...) methods into its own file.
