+++ samcli/commands/check/bottle_necks.py
@@ -3,16 +3,18 @@ Bottle neck questions are asked here. Data is saved in graph, but not calcualted
 """
 import click
 
-from samcli.commands.check.resources.Graph import Graph
-from samcli.commands.check.resources.LambdaFunction import LambdaFunction
+from samcli.commands.check.resources.graph import CheckGraph
+from samcli.commands.check.resources.lambda_function import LambdaFunction
 from samcli.commands._utils.resources import AWS_LAMBDA_FUNCTION
+
+from samcli.commands.check.lib.ask_question import ask
 
 
 class BottleNecks:
-    _graph: Graph
+    _graph: CheckGraph
     _lambda_max_duration: int
 
-    def __init__(self, graph: Graph):
+    def __init__(self, graph: CheckGraph):
         """
         Args:
             graph (Graph): The graph object. This is where all of the data is stored
@@ -40,7 +42,7 @@ class BottleNecks:
 
             entry_point_question += "\nWhere should the simulation start?"
 
-            user_input = _ask(entry_point_question, 1, item_number + 1)
+            user_input = ask(entry_point_question, 1, item_number + 1)
             current_entry_point = entry_points.pop(user_input - 1)
 
             self._ask_bottle_neck_questions(current_entry_point)
@@ -58,12 +60,12 @@ class BottleNecks:
         # If there is no entry point to the lambda function, get tps
         if lambda_function.tps == -1:
 
-            user_input_tps = _ask(
+            user_input_tps = ask(
                 "What is the expected per-second arrival rate for [%s]?\n[TPS]" % (lambda_function.resource_name)
             )
             lambda_function.tps = user_input_tps
 
-        user_input_duration = _ask(
+        user_input_duration = ask(
             "What is the expected duration for the Lambda function [%s] in ms?\n[1 - %i]"
             % (lambda_function.resource_name, self._lambda_max_duration),
             1,
@@ -81,26 +83,3 @@ class BottleNecks:
         """
         if resource.resource_type == AWS_LAMBDA_FUNCTION:
             self._lambda_bottle_neck_quesitons(resource)
-
-
-def _ask(question: str, min_val: int = 1, max_val: float = float("inf")) -> int:
-    """Prompt the user for input based on provided range
-
-    Args:
-        question (str): The question to prompt the user with
-        min_val (int, optional): Minimum acceptable value. Defaults to 1.
-        max_val (float, optional): Maximum acceptable value. Defaults to float("inf").
-
-    Returns:
-        int: User selected value
-    """
-    valid_user_input = False
-    user_input = 0
-    while not valid_user_input:
-        user_input = click.prompt(text=question, type=int)
-        if user_input > max_val or user_input < min_val:
-            click.echo("Please enter a number within the range")
-        else:
-            valid_user_input = True
-
-    return user_input
