+++ samcli/commands/check/resources/Graph.py
@@ -1,30 +1,45 @@
 """
 Class for graph. All data is stored in the graph directly, or within nodes that are stored in the graph
 """
-from typing import List
+from typing import List, Optional, Union
 
-from samcli.commands.check.resources.LambdaFunctionPricing import LambdaFunctionPricing
+from samcli.commands.check.resources.lambda_function_pricing import LambdaFunctionPricing
+from samcli.commands.check.resources.lambda_function import LambdaFunction
+from samcli.commands.check.resources.warning import CheckWarning
 
 
-class Graph:
-    entry_points: List
-    resources_to_analyze: List
-    green_warnings: List
-    yellow_warnings: List
-    red_warnings: List
-    red_burst_warnings: List
-    lambda_function_pricing_info: LambdaFunctionPricing
+class CheckGraph:
+    """
+    A graph object stores four main item types.
+    - entry_points are all possible entry points to the application based on the template.
+    - resources_to_analyze are all resources that need to be directly checked for bottle
+      neck issues (just lambda functions for now).
+    - green/yellow/red/red_burst_warnings are the four different warning types that
+      the user can be presented
+    - lambda_function_pricing_info will contain the user entered data for the lambda
+      pricing questions
+    """
 
-    def __init__(self):
-        self.entry_points: List = []
-        self.resources_to_analyze: List = []
-        self.green_warnings: List = []
-        self.yellow_warnings: List = []
-        self.red_warnings: List = []
-        self.red_burst_warnings: List = []
-        self.lambda_function_pricing_info: LambdaFunctionPricing = None
+    entry_points: List[Union[LambdaFunction]]
+    resources_to_analyze: List[Union[LambdaFunction]]
+    green_warnings: List[CheckWarning]
+    yellow_warnings: List[CheckWarning]
+    red_warnings: List[CheckWarning]
+    red_burst_warnings: List[CheckWarning]
+    lambda_function_pricing_info: Optional[LambdaFunctionPricing]
 
-    def generate(self, lambda_functions: List) -> None:
+    def __init__(self, lambda_functions: List):
+        self.entry_points: List[Union[LambdaFunction]] = []
+        self.resources_to_analyze: List[Union[LambdaFunction]] = []
+        self.green_warnings: List[CheckWarning] = []
+        self.yellow_warnings: List[CheckWarning] = []
+        self.red_warnings: List[CheckWarning] = []
+        self.red_burst_warnings: List[CheckWarning] = []
+        self.lambda_function_pricing_info: Optional[LambdaFunctionPricing] = None
+
+        self._generate(lambda_functions)
+
+    def _generate(self, lambda_functions: List) -> None:
         """Generates the graph based on the connections calulated
         Args:
             lambda_functions (List): List of all lambda functions in template
