+++ samcli/commands/check/resources/lambda_function.py
@@ -2,9 +2,9 @@
 Class object for Lambda Functions. Contains object from template, as well as all data for lambda functions,
 excluding pricing info
 """
-from typing import List, Any
+from typing import List, Any, Dict, Union
+from samcli.commands.check.resources.template_resource import TemplateResource
 from samcli.lib.providers.provider import Function
-from samcli.commands.check.resources.TemplateResource import TemplateResource
 
 
 class LambdaFunction(TemplateResource):
@@ -17,10 +17,10 @@ class LambdaFunction(TemplateResource):
     allocated_memory: int
     allocated_memory_unit: str
 
-    def __init__(self, resource_object: Function, resource_type: str):
+    def __init__(self, resource_object: Union[Dict, Function], resource_type: str):
         """
         Args:
-            resource_object (Function): The resource object form the template file
+            resource_object (Dict): The resource object form the template file
             resource_type (str): The resource type
         """
         super().__init__(resource_object, resource_type)
