+++ samcli/commands/check/results.py
@@ -4,13 +4,13 @@ Prints the results of bottle neck calculations
 from typing import List
 import click
 
-from samcli.commands.check.resources.Graph import Graph
+from samcli.commands.check.resources.graph import CheckGraph
 
 
 class Results:
-    _graph: Graph
+    _graph: CheckGraph
 
-    def __init__(self, graph: Graph):
+    def __init__(self, graph: CheckGraph):
         """
         Args:
             graph (Graph): The graph object. This is where all of the data is stored
