+++ tests/unit/commands/check/test_bottle_neck.py
@@ -2,21 +2,12 @@ from unittest import TestCase
 from unittest.mock import Mock, patch
 
 
-from samcli.commands.check.bottle_necks import BottleNecks, _ask
+from samcli.commands.check.bottle_necks import BottleNecks
 from samcli.commands._utils.resources import AWS_LAMBDA_FUNCTION
 
 
 class TestBottleNeck(TestCase):
-    @patch("samcli.commands.check.bottle_necks.click")
-    def test_ask(self, click_patch):
-        click_patch.prompt.return_value = 5
-        question = "question"
-        result = _ask(question, 1, 10)
-
-        self.assertEqual(result, click_patch.prompt.return_value)
-        click_patch.prompt.assert_called_with(text=question, type=int)
-
-    @patch("samcli.commands.check.bottle_necks._ask")
+    @patch("samcli.commands.check.bottle_necks.ask")
     @patch("samcli.commands.check.bottle_necks.click")
     def test_ask_entry_point_question(self, click_patch, patch_ask):
         graph_mock = Mock()
@@ -54,7 +45,7 @@ class TestBottleNeck(TestCase):
 
         bottle_neck._lambda_bottle_neck_quesitons.assert_called_once_with(my_resource)
 
-    @patch("samcli.commands.check.bottle_necks._ask")
+    @patch("samcli.commands.check.bottle_necks.ask")
     def test__lambda_bottle_neck_quesitons(self, patch_ask):
         lambda_function_mock = Mock()
         lambda_function_mock.tps = -1
